Compiling ADC from organized file structure
